+++ src/main/java/org/opensearch/knn/index/engine/faiss/FaissCodecFormatResolver.java
@@ -52,7 +52,7 @@ public class FaissCodecFormatResolver implements CodecFormatResolver {
         int defaultBeamWidth
     ) {
         if (isSQOneBitEncoder(params)) {
-            return new Faiss1040ScalarQuantizedKnnVectorsFormat(nativeIndexBuildStrategyFactory);
+            return new Faiss1040ScalarQuantizedKnnVectorsFormat(getApproximateThresholdValue(), nativeIndexBuildStrategyFactory);
         }
         return resolve();
     }